Avoid raising TRY200 violations within new scopes
