+++ crates/ruff/src/rules/tryceratops/rules/reraise_no_cause.rs
@@ -1,8 +1,9 @@
-use ruff_macros::{define_violation, derive_message_formats};
-use rustpython_parser::ast::{ExprKind, Stmt, StmtKind};
+use rustpython_parser::ast::{ExprKind, Stmt};
 
-use crate::ast::types::Range;
-use crate::ast::visitor::{self, Visitor};
+use ruff_macros::{define_violation, derive_message_formats};
+
+use crate::ast::helpers::RaiseStatementVisitor;
+use crate::ast::visitor::Visitor;
 use crate::checkers::ast::Checker;
 use crate::registry::Diagnostic;
 use crate::violation::Violation;
@@ -17,23 +18,6 @@ impl Violation for ReraiseNoCause {
     }
 }
 
-#[derive(Default)]
-struct RaiseStatementVisitor<'a> {
-    raises: Vec<&'a Stmt>,
-}
-
-impl<'a, 'b> Visitor<'b> for RaiseStatementVisitor<'a>
-where
-    'b: 'a,
-{
-    fn visit_stmt(&mut self, stmt: &'b Stmt) {
-        match stmt.node {
-            StmtKind::Raise { .. } => self.raises.push(stmt),
-            _ => visitor::walk_stmt(self, stmt),
-        }
-    }
-}
-
 /// TRY200
 pub fn reraise_no_cause(checker: &mut Checker, body: &[Stmt]) {
     let raises = {
@@ -44,17 +28,11 @@ pub fn reraise_no_cause(checker: &mut Checker, body: &[Stmt]) {
         visitor.raises
     };
 
-    for stmt in raises {
-        if let StmtKind::Raise { exc, cause, .. } = &stmt.node {
-            if exc
-                .as_ref()
-                .map_or(false, |expr| matches!(expr.node, ExprKind::Call { .. }))
-                && cause.is_none()
-            {
-                checker
-                    .diagnostics
-                    .push(Diagnostic::new(ReraiseNoCause, Range::from_located(stmt)));
-            }
+    for (range, exc, cause) in raises {
+        if exc.map_or(false, |expr| matches!(expr.node, ExprKind::Call { .. })) && cause.is_none() {
+            checker
+                .diagnostics
+                .push(Diagnostic::new(ReraiseNoCause, range));
         }
     }
 }
